refactor(osv): Move OSV querying to a separate module
Moves the stateless OSV querying logic from the  directory to a new  directory.

This change better reflects the nature of the code, as it is a stateless utility function rather than a stateful service. The file is renamed from  to , and all relevant import statements have been updated accordingly. No functional changes are included in this commit.
